suites upd bunch of tests
